requirements.txt
alembic==1.2.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
config==0.4.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.14.0
Mako==1.1.0
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
# google-cloud-storage==1.13.2